Split information about circuit hops into inbound and outbound.
Previously the circuit object owned not only the outbound crypto,
but also the inbound crypto and the stream maps.  That's not so
great, since the reactor needs to use the inbound crypto and the
stream maps all the time, whereas the circuit doesn't need them much
(or at all).

Moving these objects to the reactor-owned structure should let us
fix the deadlock case in stream sendme handling, since the circuit
reactor no longer needs to lock the circuit in order to do crypto
and demultiplexing.  It should also speed up the code a bit, since
it doesn't need to grab the circuit lock nearly so often as before.

This change forced me to add a couple of new reactor CtrlMsg values,
since the circuit can no longer add streams and layers directly.  I
think it will still be a performance win, though.