Improve handling of circuit closure or failure on reactor shutdown.
We already handled the case okay when we were reading on streams,
since the reactor's going away would drop the sender side of their
mpsc channels.  But if the reactor went away, nothing would tell
_writing_ streams that they needed to close.

Now we handle that case, as well as anybody who is waiting on
a meta-cell to get back to them.