Checkpoint work on a sqlite backend.
Among other stuff, revert parts of previous commit that introduced a
ReadableStore abstraction: there isn't a good abstraction hat covers
both Tor legacy stores and sqlite stores-- or if there is, I'll have
to find it organizally.

This isn't a final version: it probably has bugs, and it needs
documentation and more tests.  But before I do those, I want to
sanity-check the API by drafting the next layer up on the stack.