Light edits on CONTRIBUTING.md
Mainly, drop the suggestion to fetch submodules, and explain what
more things do.

(The only submodule we have now is the fuzzing corpora, which most
people won't want.)

(I am not 100% convinced that we should keep 100% of the
instructions for working with git in the long term.  Do many other
projects have these?  Is there a tutorial we should link to
instead?)