Periodically launch timeout-testing circuits.
Our path-spec says that, if we don't have enough build-time data to
make a good estimate for our timeouts, we should periodically try
launching testing circuits just to see how long they take to build.

Here we implement that.  The implementation is a bit wonky because
of how our crates are factored; see arti#161 for more info.

Closes #145.

Closes #57.